[chore] #82 Change folder structure in mypage
