+++ feature/mypage/src/main/java/com/cmc/curtaincall/feature/mypage/MyPageNavGraph.kt
@@ -6,20 +6,18 @@ import androidx.navigation.compose.composable
 import androidx.navigation.navigation
 import com.cmc.curtaincall.core.base.BottomDestination
 import com.cmc.curtaincall.core.base.CurtainCallDestination
-import com.cmc.curtaincall.feature.mypage.ui.MyPageScreen
-import com.cmc.curtaincall.feature.mypage.ui.ParticipantScreen
-import com.cmc.curtaincall.feature.mypage.ui.RecruitmentScreen
 import com.cmc.curtaincall.common.design.R
-import com.cmc.curtaincall.feature.mypage.ui.MyPageProfileEditScreen
+import com.cmc.curtaincall.feature.mypage.editprofile.MyPageProfileEditScreen
+import com.cmc.curtaincall.feature.mypage.saveperformance.MyPageSavedPerformance
 
 private const val MYPAGE_GRAPH = "mypage_graph"
 const val MYPAGE = "mypage"
 private const val MYPAGE_LABEL = "MY"
 private const val MYPAGE_PROFILE_EDIT = "mypage_profile_edit"
+private const val MYPAGE_SAVE_PERFORMANCE = "mypage_save_performance"
 private const val MYPAGE_RECRUITMENT = "mypage_recruitment"
 private const val MYPAGE_PARTICIPANT = "mypage_participant"
 private const val MYPAGE_WRITE = "mypage_write"
-private const val MYPAGE_SAVE = "mypage_save"
 private const val MYPAGE_ANNOUNCEMENT = "mypage_announcement"
 private const val MYPAGE_CONTACT = "mypage_contact"
 private const val MYPAGE_LOGOUT = "mypage_logout"
@@ -36,6 +34,10 @@ sealed interface MyPageDestination : CurtainCallDestination {
         override val route = MYPAGE_PROFILE_EDIT
     }
 
+    object SavePerformance : MyPageDestination {
+        override val route = MYPAGE_SAVE_PERFORMANCE
+    }
+
     object Recruitment : MyPageDestination {
         override val route = MYPAGE_RECRUITMENT
     }
@@ -46,10 +48,6 @@ sealed interface MyPageDestination : CurtainCallDestination {
 
     object Write : MyPageDestination {
         override val route = MYPAGE_WRITE
-    }
-
-    object Save : MyPageDestination {
-        override val route = MYPAGE_SAVE
     }
 
     object Announcement : MyPageDestination {
@@ -77,6 +75,14 @@ fun NavGraphBuilder.mypageNavGraph(navHostController: NavHostController) {
 
         composable(MyPageDestination.ProfileEdit.route) {
             MyPageProfileEditScreen(
+                onBack = {
+                    navHostController.popBackStack()
+                }
+            )
+        }
+
+        composable(MyPageDestination.SavePerformance.route) {
+            MyPageSavedPerformance(
                 onBack = {
                     navHostController.popBackStack()
                 }
